Update and rename B7568.java to B7568_huipyo.java
